requirements.txt
Django==5.2.5
djangorestframework==3.15.2
python-decouple==3.8
